Improved shader organization
