+++ src/shaders/links.js
@@ -1,0 +1,39 @@
+const links = {
+  vertexShader: `
+    #include <fog_pars_vertex>
+
+    uniform float is2D;
+    uniform sampler2D texturePositions;
+
+    varying vec3 vColor;
+
+    void main() {
+
+      vec3 vPosition = texture2D( texturePositions, position.xy ).xyz;
+      vPosition.z *= 1.0 - is2D;
+
+      vec4 mvPosition = modelViewMatrix * vec4( vPosition, 1.0 );
+      vColor = color;
+
+      gl_Position = projectionMatrix * mvPosition;
+      #include <fog_vertex>
+
+    }
+  `,
+  fragmentShader: `
+    #include <fog_pars_fragment>
+
+    uniform float inheritColors;
+    uniform vec3 uColor;
+    uniform float opacity;
+
+    varying vec3 vColor;
+
+    void main() {
+      gl_FragColor = vec4( mix( vec3( 1.0 ), vColor, inheritColors ) * uColor, opacity );
+      #include <fog_fragment>
+    }
+  `
+};
+
+export default links;
