+++ src/shaders/positions.js
@@ -1,0 +1,21 @@
+/**
+ * Calculate the next frame's position for all nodes.
+ */
+export const positions = `
+  uniform float is2D;
+  uniform float timeStep;
+
+  void main() {
+
+    vec2 uv = gl_FragCoord.xy / resolution.xy;
+    vec4 texel = texture2D( texturePositions, uv );
+    vec3 position = texel.xyz;
+    vec3 velocity = texture2D( textureVelocities, uv ).xyz;
+    float isStatic = texel.w;
+
+    vec3 result = position + velocity * timeStep * ( 1.0 - isStatic );
+
+    gl_FragColor = vec4( result.xyz, isStatic );
+
+  }
+`;
