+++ src/shaders/simulation.js
@@ -1,0 +1,10 @@
+import { positions } from "./positions.js";
+import { simplex, nested, types } from "./velocities.js";
+
+export default {
+  positions,
+  velocities: simplex,
+  simplex,
+  nested,
+  types
+};
